feat(w42): Week 42 Days 1-2 — WASM gating, Miri UB fix, fuzz target fixes, CI updates
- Gate wasm module behind cfg(target_arch = "wasm32") for Miri compatibility
- Move WASM deps to target-specific section in Cargo.toml
- Fix UB: replace bytemuck::from_bytes with pod_read_unaligned in persistence headers
- Fix fuzz targets: graph_ops (soft_delete API), sparse_storage (SparseId type), persistence
- Gate 11 WASM integration tests with cfg(target_arch = "wasm32")
- Add Miri CI job and fuzz-check for all targets
- Fix clippy redundant_closure in snapshot.rs
- Fix deprecated UnalignedBuffer variant in persistence_corruption test
- Add LangChain.js spike research doc (GO decision)
- Add API docs, guides, planning docs, and Miri audit review

+++ src/flat/mod.rs
@@ -573,8 +573,8 @@ mod tests {
         // Data must still be intact
         assert_eq!(index.len(), 10);
         assert_eq!(index.vectors_len(), 10 * 128);
-        for i in 1..=10 {
-            assert_eq!(index.get(VectorId(i as u64)), Some([0xAA; 128].as_slice()));
+        for i in 1..=10_u64 {
+            assert_eq!(index.get(VectorId(i)), Some([0xAA; 128].as_slice()));
         }
 
         // Shrink on empty index should not panic
